Allow the project to be built standalone
* Define all CMake macros used
* Add headerfiles to targets to make the visual
  studio projects easier to navigate in

This information is used by the project settings

Reviewed-on: http://review.couchbase.org/33929
Tested-by: Trond Norbye <trond.norbye@gmail.com>

+++ tests/getopt_test.cc
@@ -12,14 +12,14 @@
 
 char **vec2array(const std::vector<std::string> &vec) {
     char **arr = new char*[vec.size()];
-    for (int ii = 0; ii < vec.size(); ++ii) {
+    for (unsigned int ii = 0; ii < (unsigned int)vec.size(); ++ii) {
         arr[ii] = strdup(vec[ii].c_str());
     }
     return arr;
 }
 
 static void release(char **arr, size_t size) {
-    for (int ii = 0; ii < size; ++ii) {
+    for (size_t ii = 0; ii < size; ++ii) {
         free(arr[ii]);
     }
     delete []arr;
